remove requirement for user to restate ingredients
originally there was this validation function where after stating
conflicts, you would have to manually repeat every unique thing in
your skincare routine. I found this redundant, so I'm removing it.

The issue now is . . . spellchecking is no longer included. The
one benefit to the feature was that if you misspelled something in
the conflict dictionary, the validation function would alert you.

I may bring it back, but we'll see.